refactor(#3103): extract sparse Array.prototype fast paths from runtime.ts
Another bounded, byte-identical slice of the src/runtime.ts (#3103)
decomposition. Lift the #1234 sparse-aware
Array.prototype.{unshift,reverse,forEach} fast paths (for non-Array
receivers) verbatim into a new sibling module
src/runtime/array-proto-sparse.ts.

runtime.ts shrinks -187 LOC. The whole #1234 cluster references nothing
outside itself (only JS globals: Reflect/Number/Set/String/TypeError).
_collectIntegerKeys + the 3 method impls + the _arrayProtoSparseFastPaths
dispatch map move as one unit; only _arrayProtoSparseFastPaths is used by
runtime.ts (one call site in __proto_method_call) — exported and imported
back. One-directional import, no cycle. 0 stray refs to the privatized
symbols remain.

REFACTOR, zero behavior change. runtime.ts is host-side JS, not in the Wasm
emit path, so the split cannot change an emitted byte by construction.
tsc --noEmit clean; check:loc-budget green.

✓
